requirements.txt
asgiref==3.8.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.2.14
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.4
PyJWT==2.9.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
typing-extensions==4.12.2
tzdata==2022.7
urllib3==1.26.13
